+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add subscription identifiers in subscriptions & incoming publishes
+- Add message expiry interval to outgoing and incoming publications
 
 ## 0.4.1 - 2026-01-06